Big Sur: fix WindowServer crash on startup
A new function amdMtl_CreateTargetDeviceArchitecture in AMDMTLBronzeDriver has a bug where if the deviceID
is not found in a list of supported ids, then amdMtl_HWL_GetDeviceArchGroup returns NULL and crashes the
process (WindowServer). We patch AMDRadeonX4000_AMDAccelDevice::getHardwareInfo which returns the deviceID
to return a fake ID and prevent the crash.